// src/public-api/model/updateUserDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateUserDto { 
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phoneNumber?: string | null;
    isActive?: boolean;
    profilePictureUrl?: string | null;
    dateOfBirth?: string | null;
    gender?: string | null;
}